+++ tests/TestCase/Command/Install/CopyConfigCommandTest.php
@@ -36,7 +36,7 @@ class CopyConfigCommandTest extends TestCase
         $this->exec('me_cms.copy_config -v');
         $this->assertExitWithSuccess();
         foreach (Configure::read('CONFIG_FILES') as $file) {
-            $this->assertOutputContains('File or directory `' . (new Filesystem())->rtr(CONFIG . pluginSplit($file)[1] . '.php') . '` already exists');
+            $this->assertOutputContains('File or directory `' . Filesystem::instance()->rtr(CONFIG . pluginSplit($file)[1] . '.php') . '` already exists');
         }
     }
 }