+++ tests/TestCase/View/Helper/MenuHelperTest.php
@@ -85,23 +85,6 @@ class MenuHelperTest extends MenuHelperTestCase
     }
 
     /**
-     * Tests for `backups()` method
-     * @test
-     */
-    public function testBackups(): void
-    {
-        $this->assertEmpty($this->Helper->backups());
-
-        $this->writeAuthOnSession(['group' => ['name' => 'manager']]);
-        $this->assertEmpty($this->Helper->backups());
-
-        $this->writeAuthOnSession(['group' => ['name' => 'admin']]);
-        [$links,,, $handledControllers] = $this->Helper->backups();
-        $this->assertNotEmpty($links);
-        $this->assertEquals(['Backups'], $handledControllers);
-    }
-
-    /**
      * Tests for `systems()` method
      * @test
      */
@@ -109,14 +92,11 @@ class MenuHelperTest extends MenuHelperTestCase
     {
         $this->assertEmpty($this->Helper->systems());
 
-        $this->writeAuthOnSession(['group' => ['name' => 'manager']]);
-        [$links,,, $handledControllers] = $this->Helper->systems();
-        $this->assertNotEmpty($links);
-        $this->assertTextNotContains('Log management', $links);
-        $this->assertEquals(['Logs', 'Systems'], $handledControllers);
-
-        $this->writeAuthOnSession(['group' => ['name' => 'admin']]);
-        [$links] = $this->Helper->systems();
-        $this->assertTextContains('Log management', $links);
+        foreach (['manager', 'admin'] as $name) {
+            $this->writeAuthOnSession(['group' => compact('name')]);
+            [$links,,, $handledControllers] = $this->Helper->systems();
+            $this->assertNotEmpty($links);
+            $this->assertEquals(['Systems'], $handledControllers);
+        }
     }
 }