+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # 2.x branch
 ## 2.30 branch
+### 2.30.3
+* removed the files that interacted with the `cakephp-database-backup` plugin.
+    Now that is a new plugin, `me-cms-database-backup`;
+* removed `mirko-pagliai/cakephp-entity-file-log` and files that interacted with
+    log files;
+* fixed a small bug in url with query string and final slash;
+* uses sql auto-quoting to make tests more compatible;
+* added `TestCase::isMySql()` method. This allows to skip tests with different
+    schema;
+* migration to github actions.
+
 ### 2.30.2
 * fixed a small bug in the localization of static pages.